+++ tests/test_reproducibility.py
@@ -41,7 +41,7 @@ import numpy as np
 import pytest
 
 from impulse.resume import check_for_checkpoint
-from impulse.rjmcmc import RJMCMCProductSpace
+from impulse.rjmcmc import BirthDeathProductSpace
 from impulse.rjpt_sampler import RJPTSampler
 from impulse.samplers import PTSampler
 
@@ -114,7 +114,7 @@ def _rj_loglike(params):
 
 
 def _make_rj_space():
-    return RJMCMCProductSpace(
+    return BirthDeathProductSpace(
         loglikelihood=_rj_loglike,
         logprior=_rj_logprior,
         num_sources=MAX_SOURCES,